requirements.txt
aioredis==1.3.1
asgiref==3.2.3
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-crispy-forms==1.9.0
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
Pillow==7.0.0
pycparser==2.19
PyHamcrest==2.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
whitenoise==5.0.1
zope.interface==4.7.1